+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## 0.1.3
+ * Query multiple elements
+ * Functions to work with driver's system elements.
+
 ## 0.1.2
  * Documentation added